ecommerce/pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = ["Seungho <sh007.jeong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = {version = ">=4.2,<5.0", extras = ["argon2"]}
djangorestframework = ">=3.14,<4.0"
dj-database-url = ">=2.1,<3.0"
psycopg2-binary = ">=2.9,<3.0"
celery = {extras = ["redis"], version = "^5.3.6"}
argon2-cffi = "^23.1.0"
coverage = "^7.4.0"


[tool.poetry.group.dev.dependencies]
ipython = ">=8.12.0,<8.13.0"
django-extensions = ">=3.2,<4.0"
django-debug-toolbar = ">=4.2,<5.0"
black = ">=23.12,<24.0"
python-dotenv = ">=1.0,<2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
